+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Custom About window with version info and links (Website, GitHub, Documentation)
 - Build script and C bridge module for libmongoc (MongoDB C Driver) static libraries
+- MongoDB database type with BSON type mapping
 
 ## [0.9.2] - 2026-02-28